kata-webhook: Add a simple admission controller for Kata
Given that crio no longer has support for default workload trust and
containerd never had the notion of default workload trust, an user will
need to explicitly tag workloads with a runtime class for them to be
deployed with Kata. This is something most users will not do (or for
that matter should be able to choose).

A cluster operator normally defines policy based on RBAC and namespaces.

To show how Kata can be setup seamlessly by a cluster operator it makes
sense to publish a simple policy implementation using an admission
controller.

This admission controller is also needed in order to run kubernetes
conformance tests with Kata. This will allow the conformance tests
to run on all pods which can be supported with Kata. The admission
controller will exclude workloads which cannot be run in kata, while
running the bulk of the tests in Kata containers. This will give us
significantly higher kubernetes integration testing, specially as
the conformance test suite evolves over time.

Fixes: #1102
Fixes: https://github.com/kata-containers/runtime/issues/1102

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>